Extract clone for workspace to agent pkg
